refactor(plateform): clean
