+++ plateform/app/components/layout/modal.tsx
@@ -7,11 +7,10 @@ interface ModalProps extends React.HTMLAttributes<HTMLDivElement> {
   onClose: () => void;
   title: string;
   beforeTitle?: React.ReactNode;
-  showTitleIcon?: boolean;
-  titleIconClass?: string;
+  titleIcon?: string;
 }
 
-export default function Modal({ open, children, onClose, title, beforeTitle, showTitleIcon = true, titleIconClass = "fr-icon-arrow-right-line", ...props }: ModalProps) {
+export default function Modal({ open, children, onClose, title, beforeTitle, titleIcon, ...props }: ModalProps) {
   const titleId = useId();
   const [mounted, setMounted] = useState(false);
 
@@ -46,7 +45,7 @@ export default function Modal({ open, children, onClose, title, beforeTitle, sho
               <div className="fr-modal__content">
                 {beforeTitle}
                 <h2 id={titleId} className="fr-modal__title">
-                  {showTitleIcon && <span className={`${titleIconClass} fr-icon--lg`} aria-hidden="true" />}
+                  {titleIcon && <span className={`${titleIcon} fr-icon--lg`} aria-hidden="true" />}
                   {title}
                 </h2>
                 {children}
